Добавлено в README файл описание проекта, удален плагин application

+++ app/src/main/java/hexlet/code/schemas/NumberSchema.java
@@ -17,7 +17,7 @@ public final class NumberSchema extends BaseSchema<Integer> {
 
     public NumberSchema range(int startWithNumber, int endWithNumber) {
         addValidation("range", value -> value == null || value >= startWithNumber
-                && value.doubleValue() <= endWithNumber);
+                && value <= endWithNumber);
         return this;
     }
 }
